examples/sd15/pyproject.toml
[project]
name = "sd15"
version = "0.1.1"
description = "Stable Diffusion 1.5 example (inference-only; Cozy manifest via endpoint.toml)"
requires-python = ">=3.12"
dependencies = [
    "gen-worker[torch]>=0.4.3",
    "diffusers",
    "transformers<5",
    "accelerate",
    # diffusers imports some optional pipelines during module import; include
    # ftfy to avoid import-time failures in our build-time manifest discovery.
    "ftfy",
    "pillow",
    # Needed for fp8 weight-only quantization variants (diffusers TorchAoConfig).
    "torchao",
]

[dependency-groups]
dev = ["mypy>=1.10.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sd15"]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
disallow_untyped_defs = true
